+++ network/src/commonJvmAndroid/kotlin/com/wire/kalium/network/http/request/ByteChannelRequestBody.kt
@@ -1,8 +1,7 @@
-package com.wire.kalium.network
+package com.wire.kalium.network.http.request
 
 import io.ktor.utils.io.ByteReadChannel
 import io.ktor.utils.io.jvm.javaio.toInputStream
-import kotlinx.coroutines.CancellationException
 import kotlinx.coroutines.CompletableJob
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
